+++ Runtime/Utils/GraphicsUtil.cs
@@ -7,39 +7,44 @@ namespace Voxell.Graphics
 {
   public static class GraphicsUtil
   {
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static void DisposeBufferArray(ref ComputeBuffer[] bufferArray)
-    {
-      for (int b=0; b < bufferArray?.Length; b++)
-        if (bufferArray[b] != null) bufferArray[b].Dispose();
-    }
-
+    /// <summary>
+    /// Checks if compute buffers in the array has been created or not before releasing it
+    /// </summary>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
     public static void ReleaseBufferArray(ref ComputeBuffer[] bufferArray)
     {
       for (int b=0; b < bufferArray?.Length; b++)
-        if (bufferArray[b] != null) bufferArray[b].Release();
+        bufferArray[b]?.Release();
     }
 
+    /// <summary>
+    /// Checks if textures in the array has been created or not before releasing it
+    /// </summary>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static void DisposeTextureArray(ref RenderTexture[] textureArray)
+    public static void ReleaseTextureArray<T>(ref RenderTexture[] textureArray)
     {
       for (int b=0; b < textureArray?.Length; b++)
-        if (textureArray[b] != null) textureArray[b] = null;
+        textureArray[b]?.Release();
     }
 
+    /// <summary>
+    /// Checks if objects in the array has been created or not before destroying it
+    /// </summary>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static void DisposeArray<T>(ref T[] disposableArray) where T : System.IDisposable
+    public static void DestroyArray<T>(ref T[] array) where T : Object
     {
-      for (int d=0; d < disposableArray?.Length; d++)
-        if (disposableArray[d] != null) disposableArray[d].Dispose();
+      for (int b=0; b < array?.Length; b++)
+        if (array[b] != null) GameObject.Destroy(array[b]);
     }
 
+    /// <summary>
+    /// Checks if objects in the array has been created or not before disposing it
+    /// </summary>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static void ReleaseTextureArray(ref RenderTexture[] textureArray)
+    public static void DisposeArray<T>(ref T[] array) where T : System.IDisposable
     {
-      for (int b=0; b < textureArray?.Length; b++)
-        if (textureArray[b] != null) textureArray[b].Release();
+      for (int d=0; d < array?.Length; d++)
+        array[d]?.Dispose();
     }
   }